Adjusted project structure
