+++ src/views/Game/components/GameTable/index.tsx
