+++ src/views/Game/components/GameTable/styles.ts
