+++ src/views/Game/styles.ts
